+++ odoo/songs/install/data_demo.py
@@ -24,15 +24,26 @@ def import_users(ctx):
 
 
 @anthem.log
+def import_crm_team(ctx):
+    """ Importing sales team / channel from csv """
+    load_csv(ctx, 'data/demo/crm_team.csv', 'crm.team')
+
+
+@anthem.log
 def import_location(ctx):
-    """ Importing customers from csv """
+    """ Importing stock location from csv """
     load_csv(ctx, 'data/demo/stock_location.csv', 'stock.location')
 
 
 @anthem.log
 def import_customers(ctx):
-    """ Importing stock location from csv """
+    """ Importing customers from csv """
     load_csv(ctx, 'data/demo/customers.csv', 'res.partner')
+
+
+@anthem.log
+def import_drone_type(ctx):
+    load_csv(ctx, 'data/install/drone_type.csv', 'drone.type')
 
 
 @anthem.log
@@ -44,6 +55,8 @@ def import_product(ctx):
 def main(ctx):
     """ Loading demo data """
     import_users(ctx)
+    import_crm_team(ctx)
     import_location(ctx)
     import_customers(ctx)
+    import_drone_type(ctx)
     import_product(ctx)