+++ HISTORY.rst
@@ -21,6 +21,10 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* Add the import Sales Team / Channel to demo and install songs
+* Add the import "Drone type" to demo and install songs
+
+
 **Bugfixes**
 
 **Build**